+++ src/plugins/renderers/opengl/renderer/shaderparameterpack.cpp
@@ -71,7 +71,7 @@ void ShaderParameterPack::setUniform(const int glslNameId, const UniformValue &v
 
 void ShaderParameterPack::setTexture(const int glslNameId, int uniformArrayIndex, Qt3DCore::QNodeId texId)
 {
-    for (int t=0; t<m_textures.size(); ++t) {
+    for (size_t t = 0; t < m_textures.size(); ++t) {
         if (m_textures[t].glslNameId != glslNameId || m_textures[t].uniformArrayIndex != uniformArrayIndex)
             continue;
